+++ core/src/main/java/de/markusbordihn/adaptiveperformancetweakscore/AdaptivePerformanceTweaksCore.java
@@ -19,18 +19,30 @@
 
 package de.markusbordihn.adaptiveperformancetweakscore;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.IExtensionPoint;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.network.NetworkConstants;
 
+import de.markusbordihn.adaptiveperformancetweakscore.client.TitleScreenHandler;
 import de.markusbordihn.adaptiveperformancetweakscore.debug.DebugManager;
 import de.markusbordihn.adaptiveperformancetweakscore.utils.StopModReposts;
 
 @Mod(Constants.MOD_ID)
 public class AdaptivePerformanceTweaksCore {
 
+  protected static final Logger log = LogManager.getLogger(Constants.LOG_NAME);
+
   public AdaptivePerformanceTweaksCore() {
+    final IEventBus forgeEventBus = MinecraftForge.EVENT_BUS;
+
     // Make sure the mod being absent on the other network side does not cause the client to display
     // the server as incompatible
     ModLoadingContext.get().registerExtensionPoint(IExtensionPoint.DisplayTest.class,
@@ -41,6 +53,12 @@ public class AdaptivePerformanceTweaksCore {
     DebugManager.checkForDebugLogging(Constants.LOG_NAME);
 
     StopModReposts.checkStopModReposts();
+
+    // Time measurement event.
+    DistExecutor.unsafeRunWhenOn(Dist.CLIENT, () -> () -> {
+      log.info("{} Adding time measurement event listener ...", Constants.LOG_PREFIX);
+      forgeEventBus.addListener(TitleScreenHandler::handleScreenEventInit);
+    });
   }
 
 }