+++ core/src/main/java/de/markusbordihn/adaptiveperformancetweakscore/client/TitleScreenHandler.java
@@ -17,37 +17,46 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package de.markusbordihn.minecraft.adaptiveperformancetweaksmods.server;
+package de.markusbordihn.adaptiveperformancetweakscore.client;
 
-import java.util.concurrent.TimeUnit;
+import java.lang.management.ManagementFactory;
+import java.lang.management.RuntimeMXBean;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import net.minecraft.client.gui.screens.TitleScreen;
+
 import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.event.server.ServerAboutToStartEvent;
+import net.minecraftforge.client.event.ScreenEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
 
-import de.markusbordihn.minecraft.adaptiveperformancetweaksmods.Constants;
+import de.markusbordihn.adaptiveperformancetweakscore.Constants;
 
-@EventBusSubscriber(Dist.DEDICATED_SERVER)
-public class ServerHandler {
+@EventBusSubscriber(Dist.CLIENT)
+public class TitleScreenHandler {
 
   protected static final Logger log = LogManager.getLogger(Constants.LOG_NAME);
 
-  private static final long START_TIME = System.nanoTime();
   private static boolean showed = false;
 
-  protected ServerHandler() {}
+  protected TitleScreenHandler() {}
 
   @SubscribeEvent
-  public static void handleServerAboutToStartEvent(ServerAboutToStartEvent event) {
-    if (!showed) {
-      log.info("{} ⏲ Server took about {} sec to start ...", Constants.LOG_PREFIX,
-          TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - START_TIME) / 1000f);
-      showed = true;
+  public static void handleScreenEventInit(final ScreenEvent.Init event) {
+    if (event.getScreen() instanceof TitleScreen && !showed) {
+      log.debug("Found title screen for InitScreenEvent: {}", event.getScreen());
+      logMessage();
     }
   }
+
+  public static void logMessage() {
+    RuntimeMXBean runtimeMXBean = ManagementFactory.getRuntimeMXBean();
+    log.info("{} ⏲ Client took about {} sec to start ...", Constants.LOG_PREFIX,
+        (System.currentTimeMillis() - runtimeMXBean.getStartTime()) / 1000f);
+    showed = true;
+  }
+
 
 }