+++ core/src/main/java/de/markusbordihn/adaptiveperformancetweakscore/utils/StopModReposts.java
@@ -28,6 +28,7 @@ import org.apache.logging.log4j.Logger;
 
 import cpw.mods.modlauncher.Launcher;
 import cpw.mods.modlauncher.api.IEnvironment;
+
 import de.markusbordihn.adaptiveperformancetweakscore.AdaptivePerformanceTweaksCore;
 import de.markusbordihn.adaptiveperformancetweakscore.Constants;
 
@@ -43,7 +44,7 @@ public class StopModReposts {
   private static boolean isDevEnvironment =
       version.isPresent() && version.get() != null && "MOD_DEV".equals(version.get());
 
-  private static String modFileFormatRegEx = Constants.MOD_ID + "_1.19.2-\\d.\\d.\\d.jar";
+  private static String modFileFormatRegEx = Constants.MOD_ID + "_1.19.2-\\d+.\\d+.\\d+.jar";
 
   private static Pattern expectedFilePattern = Pattern.compile(modFileFormatRegEx);
 
@@ -58,7 +59,7 @@ public class StopModReposts {
     try {
       jarFilePath = AdaptivePerformanceTweaksCore.class.getProtectionDomain().getCodeSource().getLocation()
           .toURI().getPath();
-    } catch (URISyntaxException exception) {
+    } catch (SecurityException | URISyntaxException | NullPointerException exception) {
       log.error("Unable to get jar file path: {}", exception);
     }
     if (jarFilePath == null || jarFilePath.isEmpty()) {