+++ spawn/src/main/java/de/markusbordihn/adaptiveperformancetweaksspawn/spawn/SpawnManager.java
@@ -42,15 +42,15 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
 
 import de.markusbordihn.adaptiveperformancetweakscore.CoreConstants;
-import de.markusbordihn.adaptiveperformancetweakscore.entity.EntityManager;
+import de.markusbordihn.adaptiveperformancetweakscore.entity.CoreEntityManager;
 import de.markusbordihn.adaptiveperformancetweakscore.message.WarnMessages;
 import de.markusbordihn.adaptiveperformancetweakscore.player.PlayerPosition;
 import de.markusbordihn.adaptiveperformancetweakscore.player.PlayerPositionManager;
+import de.markusbordihn.adaptiveperformancetweakscore.server.ServerLevelLoad.ServerLevelLoadLevel;
 import de.markusbordihn.adaptiveperformancetweakscore.server.ServerLevelLoad;
 import de.markusbordihn.adaptiveperformancetweakscore.server.ServerLevelLoadEvent;
 import de.markusbordihn.adaptiveperformancetweakscore.server.ServerLoadEvent;
 import de.markusbordihn.adaptiveperformancetweakscore.server.ServerManager;
-import de.markusbordihn.adaptiveperformancetweakscore.server.ServerLevelLoad.ServerLevelLoadLevel;
 import de.markusbordihn.adaptiveperformancetweaksspawn.Constants;
 import de.markusbordihn.adaptiveperformancetweaksspawn.config.CommonConfig;
 
@@ -116,6 +116,10 @@ public class SpawnManager {
         log.info("{} \u2713 Enable spawn rate control with max {} per server ...",
             Constants.LOG_PREFIX, COMMON.spawnLimitationMaxMobsPerServer.get());
       }
+    }
+
+    if (CoreConstants.CREATE_LOADED) {
+      log.warn(() -> WarnMessages.disabledOptimizationModWarning(CoreConstants.CREATE_NAME));
     }
 
     if (CoreConstants.PERFORMANT_LOADED) {
@@ -190,11 +194,23 @@ public class SpawnManager {
 
     String entityName = entity.getEncodeId();
     String levelName = level.dimension().location().toString();
-    boolean isSpawnEvent = event instanceof LivingSpawnEvent;
-    String eventType = isSpawnEvent ? "spawn" : "join";
+    boolean isLivingSpawnEvent = event instanceof LivingSpawnEvent;
+    String eventType = isLivingSpawnEvent ? "spawn" : "join";
 
-    // Pre-check for relevant entities.
-    if (!isRelevantEntity(entity, entityName, levelName)) {
+    // Pre-check for ignored dimension to avoid further checks
+    if (ignoreDimensionList.contains(levelName)) {
+      log.debug("[Ignored Dimension] Allow spawn event for {} in {} ", entity, levelName);
+      return;
+    }
+
+    // Entity instance checks to ignore specific and short living entities like projectiles.
+    if (!CoreEntityManager.isRelevantEntity(entity, entityName)) {
+      return;
+    }
+
+    // Pre-check for allowed entities to avoid expensive calculations
+    if (allowList.contains(entityName)) {
+      log.debug("[Allowed Entity] Allow spawn event for {} in {} ", entity, levelName);
       return;
     }
 
@@ -206,7 +222,7 @@ public class SpawnManager {
     // Pre-check for denied entities to avoid expensive calculations.
     if (denyList.contains(entityName)) {
       log.debug("[Denied Entity] Denied {} event for {} in {} ", eventType, entity, levelName);
-      if (isSpawnEvent) {
+      if (isLivingSpawnEvent) {
         event.setResult(Event.Result.DENY);
       } else {
         event.setCanceled(true);
@@ -243,7 +259,7 @@ public class SpawnManager {
     if (spawnLimitationLimiter > 0 && spawnLimiter++ >= spawnLimitationLimiter) {
       log.debug("[Spawn Limiter {}] Blocked {} event for {} in {}.", spawnLimitationLimiter,
           eventType, entity, levelName);
-      if (isSpawnEvent) {
+      if (isLivingSpawnEvent) {
         event.setResult(Event.Result.DENY);
       } else {
         event.setCanceled(true);
@@ -253,13 +269,13 @@ public class SpawnManager {
     }
 
     // Spawn Limitations: Max mobs per Server
-    int numberOfEntities = EntityManager.getNumberOfEntities(entityName);
+    int numberOfEntities = CoreEntityManager.getNumberOfEntities(entityName);
     int spawnLimitationMaxMobsPerServer = COMMON.spawnLimitationMaxMobsPerServer.get();
     if (spawnLimitationMaxMobsPerServer > 0
         && numberOfEntities >= spawnLimitationMaxMobsPerServer) {
       log.debug("[Spawn Limitations Server: {}] Blocked {} event for {} in {}.", numberOfEntities,
           eventType, entity, levelName);
-      if (isSpawnEvent) {
+      if (isLivingSpawnEvent) {
         event.setResult(Event.Result.DENY);
       } else {
         event.setCanceled(true);
@@ -268,13 +284,13 @@ public class SpawnManager {
     }
 
     // Spawn Limitations: Max mobs per World
-    int numberOfEntitiesPerWorld = EntityManager.getNumberOfEntities(levelName, entityName);
+    int numberOfEntitiesPerWorld = CoreEntityManager.getNumberOfEntities(levelName, entityName);
     int spawnLimitationMaxMobsPerWorld = COMMON.spawnLimitationMaxMobsPerWorld.get();
     if (spawnLimitationMaxMobsPerWorld > 0
         && numberOfEntitiesPerWorld >= spawnLimitationMaxMobsPerWorld) {
       log.debug("[Spawn Limitations World: {}] Blocked {} event for {} in {}.",
           numberOfEntitiesPerWorld, eventType, entity, levelName);
-      if (isSpawnEvent) {
+      if (isLivingSpawnEvent) {
         event.setResult(Event.Result.DENY);
       } else {
         event.setCanceled(true);
@@ -286,8 +302,8 @@ public class SpawnManager {
     int numberOfEntitiesInsideViewArea = 0;
     int spawnLimitationMaxMobsPerPlayer = COMMON.spawnLimitationMaxMobsPerPlayer.get();
     if (playersPositionsInsideViewArea != null) {
-      numberOfEntitiesInsideViewArea = EntityManager.getNumberOfEntitiesInPlayerPositions(levelName,
-          entityName, playersPositionsInsideViewArea);
+      numberOfEntitiesInsideViewArea = CoreEntityManager.getNumberOfEntitiesInPlayerPositions(
+          levelName, entityName, playersPositionsInsideViewArea);
       if (spawnLimitationMaxMobsPerPlayer > 0
           && numberOfEntitiesInsideViewArea >= spawnLimitationMaxMobsPerPlayer) {
         log.debug("[Spawn Limitations Player: {}] Blocked spawn event for {} in {}.",
@@ -315,7 +331,7 @@ public class SpawnManager {
       if (limitPerWorld > 0 && numberOfEntitiesPerWorld >= limitPerWorld * spawnFactor) {
         log.debug("[World limit] Blocked {} event for {} ({} >= {} * {}f) in {}", eventType,
             entityName, numberOfEntitiesPerWorld, limitPerWorld, spawnFactor, levelName);
-        if (isSpawnEvent) {
+        if (isLivingSpawnEvent) {
           event.setResult(Event.Result.DENY);
         } else {
           event.setCanceled(true);
@@ -363,53 +379,6 @@ public class SpawnManager {
 
     // Cache result for avoid duplicated checks.
     lastAllowedSpawnEntity = entity;
-  }
-
-  private static boolean isRelevantEntity(Entity entity, String entityName, String levelName) {
-
-    // Pre-check for ignored dimension to avoid further checks
-    if (ignoreDimensionList.contains(levelName)) {
-      log.debug("[Ignored Dimension] Allow spawn event for {} in {} ", entity, levelName);
-      return false;
-    }
-
-    // Entity instance checks to ignore specific and short living entities like projectiles.
-    if (!EntityManager.isRelevantEntity(entity)) {
-      return false;
-    }
-
-    // Skip other checks if unknown entity name
-    if (entityName == null) {
-      if (entity.isMultipartEntity() || entity.getType().toString().contains("body_part")) {
-        log.debug("[Multipart Entity] Allow spawn event for {} in {}", entity, levelName);
-      } else {
-        log.warn("[Unknown Entity] Name for spawn entity {} ({}) in {} is unknown!", entity,
-            entity.getType(), levelName);
-      }
-      return false;
-    }
-
-    // Pre-check for allowed entities to avoid expensive calculations
-    if (allowList.contains(entityName)) {
-      log.debug("[Allowed Entity] Allow spawn event for {} in {} ", entity, levelName);
-      return false;
-    }
-
-    // Ignore specific entities from other mods which are not extending the right classes or using
-    // some custom definitions which could not be easily checked.
-    if (CoreConstants.MANA_AND_ARTIFICE_LOADED
-        && entityName.equals("mana-and-artifice:residual_magic")) {
-      log.debug("[Ignored Specific Entity] Ignore spawn event for {} in {}", entity, levelName);
-      return false;
-    }
-
-    // Ignore specific entities from mods which implements their own spawn handling and logic.
-    if (CoreConstants.POKECUBE_AIO_LOADED && entityName.startsWith(CoreConstants.POKECUBE_AIO_MOD)) {
-      log.debug("[Ignored Mod Entity] Ignore spawn event for {} in {}", entity, levelName);
-      return false;
-     }
-
-    return true;
   }
 
 }